added back in scaling code, as it turned out my testing image was not set up correctly as a result a 600x600 image at 200dpi was showing up on screen at 600x600 rather than the expected 300x300.
Also removed the ICC profile from the output image it was causing the image to
bloat by 4kb, which is unnecessary because it was a default sRGB profile anyway.
So the OS should assume that it has that profile regardless.

Added support for converting CMYK images to RGB via a profile, this will work
best if the image already has a profile, as otherwise I assume RSWOP.

Currently there are some known leaks in the main conversion function, right now
that's fine as the program really doesn't work for more than a single image.
However that needs to be fixed before I add directory support.